refactor: split SelectPeopleUI in to 3 files
