+++ src/components/shared/SelectPeople/index.tsx
@@ -1,0 +1,3 @@
+export * from './SelectPeopleUI'
+export * from './PeopleInList'
+export * from './PersonInChip'
