+++ src/database/helpers/group.ts
@@ -1,0 +1,5 @@
+import { GroupRecord } from '../group'
+
+export interface Group extends GroupRecord {
+    avatar?: string
+}
