+++ src/database/index.ts
@@ -1,3 +1,4 @@
 import './migrate'
 export * from './helpers/avatar'
 export * from './helpers/person'
+export * from './helpers/group'
